MixColumns

+++ GroupProject/src/AESCipher.java
@@ -3,9 +3,9 @@ public interface AESCipher {
 
 	public void encrypt(String text, String key);
 
-	public void decrypt(int[][] cipher, String key);
+	public void decrypt(byte[][] cipher, String key);
 
-	public int[][] shiftRows(int[][] state);
+	public byte[][] shiftRows(byte[][] state);
 
-	public int[][] mixColumns(int[][] state);
+	public byte[][] mixColumns(byte[][] state);
 }
